Reestruturando o projeto e adicionando arquivos e moidificacoes perdidas,
