+++ build.gradle
@@ -1,21 +1,19 @@
-subprojects {
-    apply plugin: "java"
-    sourceCompatibility = '1.10'
-    targetCompatibility = '1.10'
+apply plugin: "java"
+sourceCompatibility = '1.10'
+targetCompatibility = '1.10'
 
-    group 'com.rmbraga.github'
-    version '1.0-SNAPSHOT'
+group 'com.rmbraga.github'
+version '1.0-SNAPSHOT'
 
-    repositories {
-        mavenCentral()
-    }
+repositories {
+    mavenCentral()
+}
 
-    dependencies {
-        testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.2")
-        testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.6.2")
-    }
+dependencies {
+    testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.2")
+    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.6.2")
+}
 
-    test {
-        useJUnitPlatform()
-    }
+test {
+    useJUnitPlatform()
 }
