--- patterns/src/main/java/com/rmbraga/behavioral/enums/Carroceria.java
@@ -1,23 +1,0 @@
-package com.rmbraga.behavioral.enums;
-
-public enum Carroceria {
-    HATCHBACK("HATCHBACK"),
-    SEDA("SEDA"),
-    CONVERSIVEL("CONVERSIVEL"),
-    SUV("SUV");
-
-    private String valor;
-
-    Carroceria(String valor) {
-        this.valor = valor;
-    }
-
-    public String getValor() {
-        return valor;
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(valor);
-    }
-}
