--- patterns/src/main/java/com/rmbraga/structural/enums/PosicaoEnum.java
@@ -1,26 +1,0 @@
-package com.rmbraga.structural.enums;
-
-public enum PosicaoEnum {
-    DESENVOLVEDOR_UX("DESENVOLVEDOR UX"),
-    DESENVOLVEDOR_DEVOPS("DESENVOLVEDOR DEVOPS"),
-    DESENVOLVEDOR_FRONTEND("DESENVOLVEDOR FRONT-END"),
-    DESENVOLVEDOR_BACKEND("DESENVOLVEDOR BACK-END"),
-    GERENTE_GERAL("GERENTE GERAL"),
-    GERENTE_FINANCEIRO("GERENTE FINANCEIRO"),
-    DIRETOR("DIRETOR");
-
-    private String valor;
-
-    PosicaoEnum(String valor) {
-        this.valor = valor;
-    }
-
-    public String getValor() {
-        return valor;
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(valor);
-    }
-}
