--- patterns/src/test/java/com/rmbraga/behavioral/builders/Tests.java
@@ -1,13 +1,0 @@
-package com.rmbraga.behavioral.builders;
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Teste da aplicação principal
- */
-public class Tests {
-    @Test
-    public void aplicacaoPrincipalTest() {
-        Aplicacao.main(new String[] {});
-    }
-}
