+++ settings.gradle
@@ -1,3 +1,1 @@
 rootProject.name = 'pt-br_design_patterns'
-
-include ":patterns"
