+++ src/main/java/com/rmbraga/behavioral/builders/enums/Montadora.java
@@ -1,0 +1,25 @@
+package com.rmbraga.behavioral.builders.enums;
+
+public enum Montadora {
+    FORD("FORD"),
+    GM("GM"),
+    VOLKSWAGEN("VOLKSWAGEN"),
+    BMW("BMW"),
+    MERCEDES("MERCEDES-BENZ"),
+    AUDI("AUDI");
+
+    private String valor;
+
+    Montadora(String valor) {
+        this.valor = valor;
+    }
+
+    public String getValor() {
+        return valor;
+    }
+
+    @Override
+    public String toString() {
+        return String.valueOf(valor);
+    }
+}
