+++ src/main/java/com/rmbraga/structural/composite/Empregado.java
