+++ src/test/java/com/rmbraga/behavioral/builders/AplicacaoTest.java
@@ -1,0 +1,13 @@
+package com.rmbraga.behavioral.builders;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Teste da aplicação principal
+ */
+public class AplicacaoTest {
+    @Test
+    public void aplicacaoTest() {
+        Aplicacao.main(new String[] {});
+    }
+}
