+++ src/test/java/com/rmbraga/behavioral/builders/BuilderCarroJ8Test.java
@@ -1,0 +1,35 @@
+package com.rmbraga.behavioral.builders;
+
+import org.junit.jupiter.api.Test;
+
+import static com.rmbraga.behavioral.builders.enums.Carroceria.SEDA;
+import static com.rmbraga.behavioral.builders.enums.Cor.PRATA;
+import static com.rmbraga.behavioral.builders.enums.Montadora.GM;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+public class BuilderCarroJ8Test {
+    @Test
+    void builderCarroJ8Test() {
+        int ano = 2015;
+        int numLugares = 5;
+
+        var carro = new BuilderCarroJ8()
+                .with(c -> {
+                    c.montadora = GM;
+                    c.carroceria = SEDA;
+                    c.ano = ano;
+                    c.cor = PRATA;
+                    c.numeroDeLugares = numLugares;
+                })
+                .buildCarro();
+
+        assertNotNull(carro);
+        assertNotNull(carro.toString());
+        assertEquals(carro.getMontadora(), GM);
+        assertEquals(carro.getCarroceria(), SEDA);
+        assertEquals(carro.getCor(), PRATA);
+        assertEquals(carro.getAno(), ano);
+        assertEquals(carro.getNumeroDeLugares(), numLugares);
+    }
+}
