+++ src/test/java/com/rmbraga/structural/composite/ComposicaoTest.java
